check for duplicate applications

+++ Zap/AppState.rb
@@ -18,8 +18,8 @@ class AppState
     
     def refresh_applications
         error = Pointer.new :object
-        sym_links = NSFileManager.defaultManager.contentsOfDirectoryAtPath(@@pow_dir, error:error)
-        self.applications = sym_links.collect do |sl|
+        
+        self.applications = symlinks_in_directoy.collect do |sl|
             applicaiton = RackApplication.new
             applicaiton.symlink = sl
             directory = NSFileManager.defaultManager.destinationOfSymbolicLinkAtPath(File.join(@@pow_dir, sl), error:error)
@@ -29,8 +29,14 @@ class AppState
         end
     end
     
+    def symlinks_in_directoy
+        error = Pointer.new :object
+        sym_links = NSFileManager.defaultManager.contentsOfDirectoryAtPath(@@pow_dir, error:error)
+    end
+    
     def link_directory(directory)
         link_name = File.basename(directory)
+        return if check_existing_symlink(link_name)
         error = Pointer.new :object
         NSFileManager.defaultManager.createSymbolicLinkAtPath(File.join(@@pow_dir, link_name), withDestinationPath:directory, error:error)
         NSApplication.sharedApplication.delegate.reload_application
@@ -42,8 +48,19 @@ class AppState
     end
     
     def rename_application(application, new_name)
+        return if check_existing_symlink(new_name)
         File.rename(File.join(@@pow_dir, application.symlink), File.join(@@pow_dir, new_name)) 
         NSApplication.sharedApplication.delegate.reload_application
+    end
+    
+    def check_existing_symlink(link_name)
+        if symlinks_in_directoy.include? link_name
+            alert = NSAlert.alertWithMessageText("An application with the name \"#{link_name}\" already exists", defaultButton:"OK", alternateButton:"", otherButton:nil, informativeTextWithFormat:"Please delete or edit the old one and try again")
+            alert.runModal
+            true
+        else
+            false
+        end
     end
     
     @@instance = self.new
